web: Move shadow-template to internal
